validate_form/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.1.13
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.3
dnspython==2.4.2
idna==3.4
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.2.4
typing_extensions==4.8.0
urllib3==2.1.0
